+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on *Keep a Changelog*, and this project adheres to *Semantic
   sequence models.
 - Added sequence ASCII/Unicode rendering for single-line typed notes.
 - Added sequence ASCII/Unicode rendering for typed sequence boxes around actor groups.
+- Added sequence ASCII/Unicode rendering for typed activation state.
 
 ## [0.6.0] - 2026-05-28